getRewardWorkflows lib function and fixing type issues
